Muchisimas modificaciones y refactorizacion de algunas funciones

+++ src/main/java/com/equivalencia/equivalenciaBE/Service/CarreraServiceImpl.java
@@ -20,7 +20,7 @@ public class CarreraServiceImpl implements CarreraService {
 	}
 
 	@Override
-	public Carrera getOne(int idCarrera) {
+	public Carrera getOne(long idCarrera) {
 		return this.carreraRepository.getOne(idCarrera);
 	}
